Removed comments, dead code and added params verification.

+++ src/controllers/deleteEventById.ts
@@ -2,19 +2,23 @@ import { Request, Response } from "express";
 import prisma from '../prisma';
 
 export const deleteEventById = async (req: Request, res: Response) => {
-    const { eventId } = req.params; // Obtener el userId desde los parámetros de la ruta
+    const { eventId } = req.params;
+
+    if (!eventId) {
+        res.status(400).json({ error: 'Missing parameters.' });
+        return
+    };
 
     try {
-        // Fetch events from the database, filtrando por userId
         const events = await prisma.event.delete({
             where: {
-                id: Number(eventId), // Convertir eventId a número
+                id: Number(eventId),
             },
         });
 
-        res.json(events); // Responder con los eventos en formato JSON
+        res.json(events);
     } catch (error) {
         console.error('Error fetching events:', error);
-        res.status(500).json({ error: 'Failed to fetch events' }); // Manejar errores
+        res.status(500).json({ error: 'Failed to fetch events' });
     }
 };
