+++ src/controllers/unsubscribeUserFromEvent.ts
@@ -2,15 +2,19 @@ import { Request, Response } from "express";
 import prisma from '../prisma';
 
 export const unsubscribeUserFromEvent = async (req: Request, res: Response) => {
-    const { userId, eventId } = req.params; // Obtener el userId desde los parámetros de la ruta
+    const { userId, eventId } = req.params;
+
+    if (!userId || !eventId) {
+        res.status(400).json({ error: 'Faltan parámetros para unsubscribeUserFromEvent' });
+        return
+    };
 
     try {
-        // Fetch events from the database, filtrando por userId
         const userEvents = await prisma.eventUser.delete({
             where: {
                 userId_eventId: {
-                    userId: Number(userId), // Convertir userId a número
-                    eventId: Number(eventId), // Convertir eventId a número
+                    userId: Number(userId),
+                    eventId: Number(eventId),
                 }
             }
         });
@@ -25,9 +29,9 @@ export const unsubscribeUserFromEvent = async (req: Request, res: Response) => {
             }
         });
 
-        res.json(userEvents); // Responder con los eventos en formato JSON
+        res.json(userEvents);
     } catch (error) {
         console.error('Error fetching events:', error);
-        res.status(500).json({ error: 'Failed to fetch events' }); // Manejar errores
+        res.status(500).json({ error: 'Failed to fetch events' });
     }
 };
